Create an art.NewHMAC function.
Our implementation instantiates an HMAC using SHA256.  Thus, as
part of our public API, we need to present our specific
instantiation of the cryptographic building blocks.

Note that we've sort of down that for the different key types,
though at the moment the key handling code is still in an
internal package.